WIP - reorg establishment form section UI
