+++ shared/src/formEstablishment/FormEstablishment.dto.ts
@@ -32,11 +32,6 @@ export type WithJobAndPhone = {
   phone: Phone;
 };
 
-export type WithFirstNameAndLastName = {
-  firstName: string;
-  lastName: string;
-};
-
 export type AdminFormEstablishmentUserRight =
   GenericFormEstablishmentUserRight<"establishment-admin"> & WithJobAndPhone;
